# specifications/openapi.json
{
  "components": {
    "schemas": {
...
      "DeletionPlanResponse": {
        "description": "What deleting a project would remove, before anything is removed.\n\n`rows` is per table rather than a total so a reviewer can sanity-check\nscale. A project reporting zero messages where hundreds were expected is\nhow a wrong identifier is caught before it is acted on.",
        "properties": {
          "knowledge_revision": {
            "title": "Knowledge Revision",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "rows": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Rows",
            "type": "object"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "name",
          "knowledge_revision",
          "rows",
          "total_rows"
        ],
        "title": "DeletionPlanResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolveAttentionRequest": {
        "properties": {
          "status": {
            "default": "resolved",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ResolveAttentionRequest",
        "type": "object"
      },
...
      "SetEvidenceRoleRequest": {
        "properties": {
          "role": {
            "minLength": 1,
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "SetEvidenceRoleRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nEngineering memory for AI product discovery.\n\n**This API has no authentication.** The MVP defers authentication, teams, and\nroles, so it is safe only behind a network boundary (ADR-0014).\n\nLong operations never hold a request open: enqueueing agent work returns `202`\nwith a durable run identifier, and the client polls the run.\n",
    "title": "KAE-Memory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete this project and everything scoped to it, in one transaction.\n\n**Irreversible.** There is no archive and nothing is retained; the\n`deletion-plan` above is the only chance to see what goes.\n\nReturns what was removed rather than an empty body, so a caller can record\nit. One project per call deliberately: a batch is one decision covering many\nprojects, and the ability to delete fifty in a request is the ability to\ndelete forty-nine by accident.",
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "workspace"
        ]
      },
      "get": {
        "description": "Return one project.",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "workspace"
        ]
      }
    },
...
    "/v1/projects/{project_id}/attention": {
      "get": {
        "description": "Return the human-attention queue. Unconfirmed extraction is not this list.",
        "operationId": "list_attention_v1_projects__project_id__attention_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Open Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttentionItemResponse"
                  },
                  "title": "Response List Attention V1 Projects  Project Id  Attention Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Attention",
        "tags": [
          "synthesis"
        ]
      },
      "post": {
        "description": "Create an attention item. Extraction must not call this.",
        "operationId": "put_attention_v1_projects__project_id__attention_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutAttentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Attention",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/attention/{item_id}/resolve": {
      "post": {
        "description": "Close an attention item. Resolved items leave the queue with provenance.",
        "operationId": "resolve_attention_v1_projects__project_id__attention__item_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAttentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Attention",
        "tags": [
          "synthesis"
        ]
      }
    },
...
    "/v1/projects/{project_id}/knowledge/{item_id}/evidence-role": {
      "post": {
        "description": "Set how an extracted row participates in reasoning. The row stays.",
        "operationId": "set_evidence_role_v1_projects__project_id__knowledge__item_id__evidence_role_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEvidenceRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Evidence Role",
        "tags": [
          "synthesis"
        ]
      }
    },
...
    "/v1/projects/{project_id}/model": {
      "get": {
        "description": "Return the synthesized project model. Empty until a synthesizer runs.",
        "operationId": "list_model_v1_projects__project_id__model_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SynthesizedObjectResponse"
                  },
                  "title": "Response List Model V1 Projects  Project Id  Model Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Model",
        "tags": [
          "synthesis"
        ]
      },
      "post": {
        "description": "Create or update a working-model object. Idempotent by identity_key.",
        "operationId": "put_model_v1_projects__project_id__model_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSynthesizedObjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesizedObjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Model",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/model/{object_id}": {
      "get": {
        "description": "Return one synthesized object and the evidence it is mapped to.",
        "operationId": "get_model_object_v1_projects__project_id__model__object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesizedObjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model Object",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/model/{object_id}/correct": {
      "post": {
        "description": "Record a person's wording as authoritative. Evidence is not deleted.",
        "operationId": "correct_model_object_v1_projects__project_id__model__object_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectSynthesizedObjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesizedObjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Model Object",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/model/{object_id}/evidence": {
      "post": {
        "description": "Map a synthesized object onto an extracted row without deleting the row.",
        "operationId": "bind_model_evidence_v1_projects__project_id__model__object_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceBindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind Model Evidence",
        "tags": [
          "synthesis"
        ]
      }
    },
...
    "/v1/projects/{project_id}/reconciliation/events": {
      "get": {
        "description": "Return recorded reconciliation/change events.",
        "operationId": "list_changes_v1_projects__project_id__reconciliation_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReconciliationEventResponse"
                  },
                  "title": "Response List Changes V1 Projects  Project Id  Reconciliation Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Changes",
        "tags": [
          "synthesis"
        ]
      },
      "post": {
        "description": "Record a change event, or return the one already stored under this key.",
        "operationId": "record_change_v1_projects__project_id__reconciliation_events_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Change",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/reconciliation/neighborhoods/{item_id}": {
      "get": {
        "description": "Return bounded related evidence for later LLM context. Not attention.",
        "operationId": "get_neighborhood_v1_projects__project_id__reconciliation_neighborhoods__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neighborhood",
        "tags": [
          "synthesis"
        ]
      }
    },
    "/v1/projects/{project_id}/reconciliation/runs": {
      "post": {
        "description": "Run deterministic evidence-graph reconciliation. Does not mint a model.",
        "operationId": "run_reconciliation_v1_projects__project_id__reconciliation_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunReconciliationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Reconciliation",
        "tags": [
          "synthesis"
        ]
      }
    },
...
  }
}